feat: JWT - 토큰 인가 에러 처리

+++ src/main/java/sigma/Spring_backend/jwt/config/Jwt.java
@@ -9,5 +9,5 @@ public interface Jwt {
 	String REFRESH_TOKEN_HEADER = "Authorization-refresh";
 	String TOKEN_PREFIX = "Bearer ";
 	int ACCESS_TOKEN_EXPIRATION = (1000 * 60) * 30; // 30min
-	int REFRESH_TOKEN_EXPIRATION = (1000 * 60 * 60 * 24) * 14; // 14day
+	long REFRESH_TOKEN_EXPIRATION = (1000 * 60 * 60 * 24) * 14; // 14day
 }
